requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
demjson==2.2.4

hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.1

numpy==1.17.2
pandas==0.25.1
parsel==1.5.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.3
queuelib==1.5.0
requests==2.22.0
Scrapy==1.7.3
scrapyrt==0.11.0
service-identity==18.1.0
six==1.12.0
Twisted==19.7.0
urllib3==1.25.6
w3lib==1.21.0
Werkzeug==0.16.0
zope.interface==4.6.0
pymongo~=3.11.2
Pillow~=6.2.1
Flask~=1.1.1
python-slugify~=4.0.1
flask-caching~=1.9.0
passlib~=1.7.4
dnspython==2.1.0
markupsafe==2.0.1

torch==1.12.1
transformers==4.21.1
tqdm==4.64.0


fastapi==0.87.0
pydantic
uvicorn


pymongo[srv]
